Centralize ball movement in xpos property
This mostly makes it so that external updates required for consistency
(e.g.  the UI_POS_NAVBALL setting) stay in one place.

I tried to obsolete the private field backing xpos, but the ConfigNode
serialization functions don't seem to like C# properties and I'm not
sure why. I found I could limit it to a stepping stool for save-load,
though.