Force ksp version max to 1.12.99
Apparently, if ksp_max_version is omitted from the packaged .version
file, and the mod is hosted on Spacedock, ckan uses the Spacedock
version. And Spacedock has no way to leave the max version open.

So I have no choice but to specify a (bogus) max version in the .version
file.  1.12.x is supposedly the last release version of KSP, so this
*shouldn't* break in future, but it still makes me react with fear and
loathing.

References:

* https://github.com/KSP-CKAN/NetKAN/issues/7166
* https://github.com/KSP-CKAN/NetKAN/issues/7697
* https://github.com/KSP-CKAN/NetKAN/issues/8333
* https://github.com/KSP-CKAN/NetKAN/pull/7729